Only reset GMS data if there is something to reset

+++ zip-content/inc/common.sh
@@ -262,9 +262,11 @@ zip_extract_dir()
 # Data reset functions
 reset_gms_data_of_all_apps()
 {
-  ui_debug 'Resetting GMS data of all apps...'
-  find /data/data/*/shared_prefs -name "com.google.android.gms.*.xml" -delete
-  validate_return_code "$?" 'Failed to reset GMS data of all apps'
+  if test -e '/data/data/'; then
+    ui_debug 'Resetting GMS data of all apps...'
+    find /data/data/*/shared_prefs -name 'com.google.android.gms.*.xml' -delete
+    validate_return_code "$?" 'Failed to reset GMS data of all apps'
+  fi
 }
 
 # Hash related functions
